Paginado y obtiene el usuario para mostrar en pantalla

+++ src/main/java/com/curso/model/Nota.java
@@ -52,7 +52,7 @@ public class Nota implements Serializable {
     private String comentarioAdm;
 
     @Column(name = "valorizacion")
-    private short valorizacion = 1;
+    private Integer valorizacion = 1;
 
     public int getCodigo() {
         return codigo;
@@ -110,11 +110,11 @@ public class Nota implements Serializable {
         this.comentarioAdm = comentarioAdm;
     }
 
-    public short getValorizacion() {
+    public Integer getValorizacion() {
         return valorizacion;
     }
 
-    public void setValorizacion(short valorizacion) {
+    public void setValorizacion(Integer valorizacion) {
         this.valorizacion = valorizacion;
     }
